changed printing function to add rounding

+++ lib/OLED/oled.h
@@ -2,11 +2,11 @@
 #include <Wire.h>
 
 class Display {
-    public:
-        void init();
-        void SetUnits(char units_str[4]);
-        void SetSpeed(double speed_inp);
-    private:
-        char units[4];
-        double speed;
+public:
+    void init();
+    void SetUnits(char units_str[4]);
+    void SetSpeed(float speed_inp);
+private:
+    char units[4];
+    double speed;
 };
